Allow STL files written by Magics Materialise to be read by DREAM.3D.
Write UTF8 based header in STL files instead of UTF16.

Fixes #732. Closes #732

+++ Source/Plugins/IO/IOFilters/ReadStlFile.h
@@ -164,12 +164,12 @@ class ReadStlFile : public AbstractFilter
   private:
     DEFINE_DATAARRAY_VARIABLE(double, FaceNormals)
 
-    double m_minXcoord;
-    double m_maxXcoord;
-    double m_minYcoord;
-    double m_maxYcoord;
-    double m_minZcoord;
-    double m_maxZcoord;
+    float m_minXcoord;
+    float m_maxXcoord;
+    float m_minYcoord;
+    float m_maxYcoord;
+    float m_minZcoord;
+    float m_maxZcoord;
 
     /**
      * @brief updateFaceInstancePointers Updates raw Face pointers
